Adds support for bluetooth serial port.
We now have an hc05 compatible module hooked to A5 and A4.

The TX pin from the module goes straight to A5.
The RX pin is connected to A4 via a resistive divider. There are many
vairations around but a 2k2 resistor to ground and a 1k to the Arduino
works fine.

Power is taken from the Vref output of the Arduino.